+++ plugins/docker/docker.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"fmt"
-	"github.com/tetra2000/canary/api"
+	"github.com/tetra2000/canary/api/types"
 	"github.com/tetra2000/canary/plugins/docker/lib"
 
-	"github.com/docker/docker/api/types"
+	dockerTypes "github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 	"context"
 )
@@ -24,7 +24,7 @@ func (p DockerPlugin) Name() string {
 	return "DockerPlugin"
 }
 
-func (p DockerPlugin) Exec(taskName string, args api.PluginArg) api.PluginResult {
+func (p DockerPlugin) Exec(taskName string, args types.PluginArg) types.PluginResult {
 	fmt.Print("Listing Docker containers.\n")
 
 	cli, err := client.NewEnvClient()
@@ -32,7 +32,7 @@ func (p DockerPlugin) Exec(taskName string, args api.PluginArg) api.PluginResult
 		panic(err)
 	}
 
-	containers, err := cli.ContainerList(context.Background(), types.ContainerListOptions{})
+	containers, err := cli.ContainerList(context.Background(), dockerTypes.ContainerListOptions{})
 	if err != nil {
 		panic(err)
 	}
@@ -41,7 +41,7 @@ func (p DockerPlugin) Exec(taskName string, args api.PluginArg) api.PluginResult
 		fmt.Printf("%s %s\n", container.ID[:10], container.Image)
 	}
 
-	return api.PluginResult{Output: "", Err: nil}
+	return types.PluginResult{Output: "", Err: nil}
 }
 
 var Plugin DockerPlugin
